refactor: create SearchResults component from Search
