+++ src/app/index.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Search } from './components/search/search';
+import { Search } from './components/search';
 import { ErrorBoundary } from './components/errorBoundary';
 
 export class App extends Component {
